Feat : Load<T>(string key) 추가, SpawnProjectile()추가
1. string Key를 토대로 obj를 뱉는 Load<T>(string key) 추가

2. 모든 Projectile타입 스킬을 스폰해주는 메소드 작성 예정

+++ Assets/01.Scripts/Skill/SkillBase.cs
@@ -63,7 +63,7 @@ public class SkillBase : MonoBehaviour
     #endregion
 
 
-    public float TotalDamage { get; set; } = 0;
+    public int TotalDamage { get; set; } = 0;
     // SetInfoSkillData 스크립터블 읽어와서 쏵 세팅해주기.
     public void UpdateSkillData()
     {
